+++ packages/core/src/index.ts
@@ -1,2 +1,2 @@
-export { getIde, getFiltersForIde } from './environment';
+export { getDevSpace, getIde, getFiltersForIde } from './environment';
 export { getGuidedAnswerApi } from './guided-answers-api';