+++ packages/ide-extension/src/enhancement/enhancements.ts
@@ -1,6 +1,6 @@
 import { extensions } from 'vscode';
 import { isVSCodeCommand } from '@sap/guided-answers-extension-types';
-import type { HTMLEnhancement, NodeEnhancement, IDE } from '@sap/guided-answers-extension-types';
+import type { HTMLEnhancement, IDE } from '@sap/guided-answers-extension-types';
 import { logString } from '../logger/logger';
 
 import enhancementsJson from './enhancements.json';
@@ -9,19 +9,19 @@ import enhancementsJson from './enhancements.json';
  * Classifies enhancements as applicable and inapplicable. There can be different reasons for a command to be
  * inapplicable, most likely because the extension hosting the command is not installed.
  *
- * @param enhancements - array of enhancements, either HTMLEnhancement or NodeEnhancement
+ * @param enhancements - array of HTMLEnhancement
  * @param ide - runtime environment like VSCODE or SBAS
  * @returns - two arrays: applicable and inapplicable extensions
  */
-function classifyEnhancements<T extends HTMLEnhancement | NodeEnhancement>(
-    enhancements: T[],
+function classifyEnhancements(
+    enhancements: HTMLEnhancement[],
     ide: IDE
-): { applicable: T[]; inapplicable: T[] } {
-    const applicable: T[] = [];
-    const inapplicable: T[] = [];
+): { applicable: HTMLEnhancement[]; inapplicable: HTMLEnhancement[] } {
+    const applicable: HTMLEnhancement[] = [];
+    const inapplicable: HTMLEnhancement[] = [];
 
     for (const enhancement of enhancements) {
-        if (enhancement.command.environment.includes(ide)) {
+        if (enhancement.command.environment && enhancement.command.environment.includes(ide)) {
             if (isVSCodeCommand(enhancement.command.exec)) {
                 if (extensions.getExtension(enhancement.command.exec.extensionId)) {
                     applicable.push(enhancement);
@@ -44,18 +44,8 @@ function classifyEnhancements<T extends HTMLEnhancement | NodeEnhancement>(
  * @param ide - runtime environment like VSCODE or SBAS
  * @returns - arrays applicable of html enhancements and node enhancements
  */
-export function getEnhancements(ide: IDE): {
-    nodeEnhancements: NodeEnhancement[];
-    htmlEnhancements: HTMLEnhancement[];
-} {
-    const htmlEnhancements = classifyEnhancements<HTMLEnhancement>(
-        enhancementsJson.htmlEnhancements as HTMLEnhancement[],
-        ide
-    );
-    const nodeEnhancements = classifyEnhancements<NodeEnhancement>(
-        enhancementsJson.nodeEnhancements as NodeEnhancement[],
-        ide
-    );
+export function getHtmlEnhancements(ide: IDE): HTMLEnhancement[] {
+    const htmlEnhancements = classifyEnhancements(enhancementsJson.htmlEnhancements as HTMLEnhancement[], ide);
 
     if (htmlEnhancements.applicable.length > 0) {
         logString(`Applicable html enhancements:\n${JSON.stringify(htmlEnhancements.applicable, null, 2)}`);
@@ -69,21 +59,14 @@ export function getEnhancements(ide: IDE): {
             )}`
         );
     }
-    if (nodeEnhancements.applicable.length > 0) {
-        logString(`Applicable html enhancements:\n${JSON.stringify(nodeEnhancements.applicable, null, 2)}`);
-    }
-    if (nodeEnhancements.inapplicable.length > 0) {
-        logString(
-            `Following node enhancements can not be applied and will not appear in Guided Answers:\n${JSON.stringify(
-                nodeEnhancements.inapplicable,
-                null,
-                2
-            )}`
-        );
-    }
+    return htmlEnhancements.applicable;
+}
 
-    return {
-        nodeEnhancements: nodeEnhancements.applicable,
-        htmlEnhancements: htmlEnhancements.applicable
-    };
+/**
+ * Get a string set of all installed extension ids.
+ *
+ * @returns - set with all installed extensions
+ */
+export function getInstalledExtensionIds(): Set<string> {
+    return new Set<string>(extensions.all.map((ext) => ext.id.toLowerCase()));
 }