+++ packages/ide-extension/test/__mocks__/@vscode/extension-telemetry.js
@@ -1,0 +1,7 @@
+module.exports = class TelemetryReporter {
+    constructor() {
+        return {
+            sendTelemetryEvent: jest.fn()
+        };
+    }
+};