+++ packages/webapp/src/webview/state/middleware.ts
@@ -1,6 +1,12 @@
 import i18next from 'i18next';
 import type { Middleware, MiddlewareAPI, Dispatch, Action } from 'redux';
 import type { GuidedAnswerActions } from '@sap/guided-answers-extension-types';
+import {
+    GO_TO_PREVIOUS_PAGE,
+    SEND_TELEMETRY,
+    SET_ACTIVE_TREE,
+    UPDATE_ACTIVE_NODE
+} from '@sap/guided-answers-extension-types';
 import type { AppState } from '../types';
 
 declare let window: Window;
@@ -19,7 +25,7 @@ export const communicationMiddleware: Middleware<
     AppState,
     Dispatch<GuidedAnswerActions>
 > = (store: MiddlewareAPI) => {
-    // Add event handler, this will dispatch incomming state updates
+    // Add event handler, this will dispatch incoming state updates
     window.addEventListener('message', (event: MessageEvent) => {
         if (event.origin === window.origin) {
             console.log(i18next.t('MESSAGE_RECEIVED'), event);
@@ -38,9 +44,29 @@ export const communicationMiddleware: Middleware<
     return (next: Dispatch<GuidedAnswerActions>) =>
         (action): Action => {
             action = next(action);
-            if (action && typeof action.type === 'string' && !action.type.startsWith('[view]')) {
+            if (action && typeof action.type === 'string') {
                 window.vscode.postMessage(action);
                 console.log(i18next.t('REACT_ACTION_POSTED'), action);
+            }
+            return action;
+        };
+};
+
+const allowedTelemetryActions = new Set([GO_TO_PREVIOUS_PAGE, SET_ACTIVE_TREE, UPDATE_ACTIVE_NODE]);
+export const telemetryMiddleware: Middleware<
+    Dispatch<GuidedAnswerActions>,
+    AppState,
+    Dispatch<GuidedAnswerActions>
+> = ({ getState }) => {
+    return (next: Dispatch<GuidedAnswerActions>) =>
+        (action: GuidedAnswerActions): GuidedAnswerActions => {
+            action = next(action);
+            if (action && typeof action.type === 'string' && allowedTelemetryActions.has(action.type)) {
+                window.vscode.postMessage({
+                    type: SEND_TELEMETRY,
+                    payload: { action, state: getState() }
+                } as unknown as JSON);
+                console.log(i18next.t('TELEMETRY_POSTED, action: '), action);
             }
             return action;
         };